+++ src/main/java/roomescape/controller/ReservationAdminController.java
@@ -1,0 +1,18 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ReservationAdminController {
+
+    @GetMapping("/admin")
+    public String homePage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String manageReservation() {
+        return "admin/reservation-legacy";
+    }
+}
